Cleanup of ci.yml
Testing github ci badge

Testing github ci badgegp

Testing github ci badgegp

fix: use data_dir from config.toml if present

fixes: https://todo.sr.ht/~gheartsfield/nostr-rs-relay/64

docs: typo in database-maintenance.md

Add a missing `"`

improvement: default to having a event creation rate limit (5/sec)

improvement: add NostrRepo trait, with sqlite implementation

This is inspired by the work of
v0l (https://github.com/v0l/nostr-rs-relay/).

A new trait abstracts the storage layer with an async API.  Rusqlite
is still used with worker threads, but this allows for Postgresql or
other backends to be used.

There may be bugs, this has not been rigorously tested.

refactor: reformat and remove tabs

feat: prometheus metrics

Prometheus metrics exposed at /metrics

fix: integration tests working

Testing github ci badgegp
