+++ .github/workflows/ci.yml
@@ -1,13 +1,11 @@
-name: CI
+name: Test and build
 
 on:
   push:
     branches:
       - master
-      - chore/add-CI
-
+      
 jobs:
-
   test_nostr-rs-relay:
     runs-on: ubuntu-latest
     steps:
@@ -24,7 +22,6 @@ jobs:
           cargo --version --verbose
           rustc --version
           cargo clippy --version          
-          ls -lsrth
 
       # - name: Lint
       #   run: |
