+++ README.md
@@ -10,6 +10,9 @@ mirrored on [GitHub](https://github.com/scsibug/nostr-rs-relay).
 
 [![builds.sr.ht status](https://builds.sr.ht/~gheartsfield/nostr-rs-relay/commits/master.svg)](https://builds.sr.ht/~gheartsfield/nostr-rs-relay/commits/master?)
 
+![example workflow](https://github.com/schlunsen/nostr-rs-relay/actions/workflows/ci.yml/badge.svg)
+
+
 ## Features
 
 [NIPs](https://github.com/nostr-protocol/nips) with a relay-specific implementation are listed here.
