+++ src/cli.rs
@@ -7,8 +7,7 @@ pub struct CLIArgs {
         short,
         long,
         help = "Use the <directory> as the location of the database",
-        default_value = ".",
-        required = false
+        required = false,
     )]
-    pub db: String,
+    pub db: Option<String>,
 }
