+++ src/lib.rs
@@ -10,7 +10,7 @@ pub mod hexrange;
 pub mod info;
 pub mod nip05;
 pub mod notice;
-pub mod schema;
+pub mod repo;
 pub mod subscription;
 pub mod utils;
 // Public API for creating relays programatically
