+++ src/utils.rs
@@ -2,7 +2,7 @@
 use std::time::SystemTime;
 
 /// Seconds since 1970.
-pub fn unix_time() -> u64 {
+#[must_use] pub fn unix_time() -> u64 {
     SystemTime::now()
         .duration_since(SystemTime::UNIX_EPOCH)
         .map(|x| x.as_secs())
@@ -10,12 +10,12 @@ pub fn unix_time() -> u64 {
 }
 
 /// Check if a string contains only hex characters.
-pub fn is_hex(s: &str) -> bool {
+#[must_use] pub fn is_hex(s: &str) -> bool {
     s.chars().all(|x| char::is_ascii_hexdigit(&x))
 }
 
 /// Check if a string contains only lower-case hex chars.
-pub fn is_lower_hex(s: &str) -> bool {
+#[must_use] pub fn is_lower_hex(s: &str) -> bool {
     s.chars().all(|x| {
         (char::is_ascii_lowercase(&x) || char::is_ascii_digit(&x)) && char::is_ascii_hexdigit(&x)
     })
